refactor: split api client modules
